+++ ui/signed-out/build.gradle.kts
@@ -4,6 +4,7 @@ plugins {
     id("com.android.library")
     id("social.androiddev.code-quality")
     id("kotlin-parcelize")
+    id("com.google.osdetector")
 }
 
 val targetSDKVersion: Int by rootProject.extra
@@ -61,6 +62,26 @@ kotlin {
             dependencies {
 //                dependsOn(commonMain)
                 implementation(compose.desktop.common)
+                //FIXME : Find the right way to use javafx in kotlin multiplatform project with android plugin
+                // https://stackoverflow.com/questions/73187027/use-javafx-in-kotlin-multiplatform
+                // As JavaFX have platform-specific dependencies, we need to add them manually
+                val fxSuffix = when (osdetector.classifier) {
+                    "linux-x86_64" -> "linux"
+                    "linux-aarch_64" -> "linux-aarch64"
+                    "windows-x86_64" -> "win"
+                    "osx-x86_64" -> "mac"
+                    "osx-aarch_64" -> "mac-aarch64"
+                    else -> throw IllegalStateException("Unknown OS: ${osdetector.classifier}")
+                }
+
+                // Replace "compileOnly" with "implementation" for a non-library project
+                implementation("org.openjfx:javafx-base:19:${fxSuffix}")
+                implementation("org.openjfx:javafx-graphics:19:${fxSuffix}")
+                implementation("org.openjfx:javafx-controls:19:${fxSuffix}")
+                implementation("org.openjfx:javafx-web:19:${fxSuffix}")
+                implementation("org.openjfx:javafx-swing:19:${fxSuffix}")
+                implementation("org.openjfx:javafx-media:19:${fxSuffix}")
+
             }
         }
     }
