+++ ui/signed-out/src/commonMain/kotlin/social/androiddev/signedout/signin/SignInViewModel.kt
@@ -7,7 +7,7 @@
  *
  * You should have received a copy of the GNU General Public License along with Dodo. If not, see <https://www.gnu.org/licenses/>.
  */
-package social.androiddev.common.composables.webview
+package social.androiddev.signedout.signin
 
 import java.net.URI
